Rename notification logic module to avoid case-only filename collision
